Cog cleanup and refactor
Moved everything into it's own cogs
