+++ src/cogs/general/general_define.py
@@ -2,11 +2,15 @@
 searches a free API: https://api.dictionaryapi.dev/api/v2/entries/en/{word}
 and returns the definition of a word
 """
+import logging
 import json
 import requests
 from discord.ext import commands
 
-from utilities.cog_helpers._embeds import embed_definition
+from utilities.cog_helpers._embeds import embed_definition  # pylint: disable=E0401
+
+
+logger = logging.getLogger(__name__)
 
 
 class Define(commands.Cog):
@@ -24,6 +28,9 @@ class Define(commands.Cog):
         make an embed
         handle the errors
         """
+        logger.info("%s used the %s command."
+                    , ctx.author.name
+                    , ctx.command)
         data = json.loads(
             requests.get(
                 f'https://api.dictionaryapi.dev/api/v2/entries/en/{word}'
