+++ src/cogs/logging/logging_avatars.py
@@ -3,8 +3,8 @@ Logs user avatar changes.
 TODO: Would be cool to add an API that detects nasty images here.
 """
 from discord.ext import commands
-from utilities.cog_helpers._embeds import embed_avatar
-from cogs._settings import log_channel
+from utilities.cog_helpers._embeds import embed_avatar  # pylint: disable=E0401
+from cogs._settings import log_channel  # pylint: disable=E0401
