+++ src/cogs/logging/logging_member_ban.py
@@ -2,8 +2,8 @@
 Logs member bans
 """
 from discord.ext import commands
-from utilities.cog_helpers._embeds import embed_ban
-from cogs._settings import log_channel, server_info
+from utilities.cog_helpers._embeds import embed_ban  # pylint: disable=E0401
+from cogs._settings import log_channel, server_info  # pylint: disable=E0401
 
 
 class LoggingBans(commands.Cog):
