+++ src/cogs/logging/logging_member_kick.py
@@ -2,8 +2,8 @@
 Logs when a user is kicked
 """
 from discord.ext import commands
-from utilities.cog_helpers._embeds import embed_kick
-from cogs._settings import log_channel, server_info
+from utilities.cog_helpers._embeds import embed_kick  # pylint: disable=E0401
+from cogs._settings import log_channel, server_info  # pylint: disable=E0401
 
 
 class LoggingKicks(commands.Cog):
