+++ src/cogs/logging/logging_member_roles.py
@@ -2,8 +2,8 @@
 Logging for role changes. Logs the user who did the changing, the target user and the role.
 """
 from discord.ext import commands
-from utilities.cog_helpers._embeds import embed_role_add, embed_role_remove
-from cogs._settings import log_channel, server_info
+from utilities.cog_helpers._embeds import embed_role_add, embed_role_remove  # pylint: disable=E0401
+from cogs._settings import log_channel, server_info  # pylint: disable=E0401
