+++ src/cogs/logging/logging_tickets.py
@@ -2,9 +2,8 @@
 logs when someone makes a ticket.
 """
 from discord.ext import commands
-from utilities.cog_helpers._embeds \
-    import embed_ticket_create, embed_ticket_update, embed_ticket_delete, embed_ticket_remove
-from cogs._settings import log_channel, server_info
+from utilities.cog_helpers._embeds import embed_ticket_create, embed_ticket_update, embed_ticket_delete, embed_ticket_remove  # pylint: disable=E0401
+from cogs._settings import log_channel, server_info  # pylint: disable=E0401
